ChatOnline/requirements.txt
Flask-MySQLdb==0.2.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
Flask-Login==0.4.1
